tests/submission/PhilipTaylor/canvas/fallback.multiple.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: fallback.multiple</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.fallback.html">fallback</a>.multiple</h1>
<p class="desc">Fallback content with multiple elements</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.fallback">fallback</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL</p><p class="fallback">FAIL</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Fallback content with multiple elements");
_addTest(function(canvas, ctx) {

_assertEqual(canvas.childNodes.length, 2, "canvas.childNodes.length", "2");


});
</script>

<div id="log"></div>
</body></html>
